Make sure things are in place to run on Cosmos successfully
