+++ src/Migration.AcceptanceTests/When_migrated_saga_completed.cs
@@ -15,6 +15,8 @@ namespace NServiceBus.AcceptanceTests
         [Test]
         public async Task Should_remove_secondary_index()
         {
+            Requires.AzureTables();
+
             var correlationPropertyValue = Guid.NewGuid();
             var sagaId = Guid.NewGuid();
