+++ src/NonTransactionalSagasWithConventionalTables.AcceptanceTests/SetupFixture.cs
@@ -1,7 +1,6 @@
 ﻿namespace NServiceBus.AcceptanceTests
 {
     using System.Linq;
-    using System;
     using System.Threading.Tasks;
     using NUnit.Framework;
     using Microsoft.Azure.Cosmos.Table;
@@ -12,7 +11,7 @@
         [OneTimeSetUp]
         public async Task OneTimeSetUp()
         {
-            var connectionString = GetEnvConfiguredConnectionStringForPersistence();
+            var connectionString = ConnectionStringHelper.GetEnvConfiguredConnectionStringForPersistence();
 
             var account = CloudStorageAccount.Parse(connectionString);
             TableClient = account.CreateCloudTableClient();
@@ -32,24 +31,6 @@
         {
             // we can't delete the tables due to conflicts when recreating
             return Task.CompletedTask;
-        }
-
-        public static string GetEnvConfiguredConnectionStringForPersistence()
-        {
-            var environmentVartiableName = "AzureStoragePersistence_ConnectionString";
-            var connectionString = GetEnvironmentVariable(environmentVartiableName);
-            if (string.IsNullOrEmpty(connectionString))
-            {
-                throw new Exception($"Oh no! We couldn't find an environment variable '{environmentVartiableName}' with Azure Storage connection string.");
-            }
-
-            return connectionString;
-        }
-
-        static string GetEnvironmentVariable(string variable)
-        {
-            var candidate = Environment.GetEnvironmentVariable(variable, EnvironmentVariableTarget.User);
-            return string.IsNullOrWhiteSpace(candidate) ? Environment.GetEnvironmentVariable(variable) : candidate;
         }
 
         public static CloudTableClient TableClient;
