+++ src/PersistenceTests/PersistenceTests.csproj
@@ -21,6 +21,10 @@
   </ItemGroup>
 
   <ItemGroup>
+    <Compile Include="..\SharedAcceptanceTests.All\ConnectionStringHelper.cs" />
+  </ItemGroup>
+
+  <ItemGroup>
     <!-- Remove once https://github.com/Particular/NServiceBus/pull/5808 is merged and new alpha released -->
     <Compile Remove="\**\Outbox\*.cs" />
   </ItemGroup>
