+++ src/Subscription.AcceptanceTests/SetupFixture.cs
@@ -12,7 +12,7 @@
         [OneTimeSetUp]
         public async Task OneTimeSetUp()
         {
-            var connectionString = GetEnvConfiguredConnectionStringForPersistence();
+            var connectionString = ConnectionStringHelper.GetEnvConfiguredConnectionStringForPersistence();
 
             TableName = $"{Path.GetFileNameWithoutExtension(Path.GetTempFileName())}{DateTime.UtcNow.Ticks}".ToLowerInvariant();
 
@@ -34,24 +34,6 @@
         public Task OneTimeTearDown()
         {
             return Table.DeleteIfExistsAsync();
-        }
-
-        public static string GetEnvConfiguredConnectionStringForPersistence()
-        {
-            var environmentVartiableName = "AzureStoragePersistence_ConnectionString";
-            var connectionString = GetEnvironmentVariable(environmentVartiableName);
-            if (string.IsNullOrEmpty(connectionString))
-            {
-                throw new Exception($"Oh no! We couldn't find an environment variable '{environmentVartiableName}' with Azure Storage connection string.");
-            }
-
-            return connectionString;
-        }
-
-        static string GetEnvironmentVariable(string variable)
-        {
-            var candidate = Environment.GetEnvironmentVariable(variable, EnvironmentVariableTarget.User);
-            return string.IsNullOrWhiteSpace(candidate) ? Environment.GetEnvironmentVariable(variable) : candidate;
         }
 
         public static string TableName;
